yeah me abdul rehman commited

+++ StudentManagementSystem.Web/Models/ExamResultDisplayViewModel.cs
@@ -4,7 +4,7 @@
     {
         public string StudentName { get; set; }
         public string CourseName { get; set; }
-        public int Marks { get; set; }
+        public double Marks { get; set; }
         public string Grade { get; set; }
     }
